backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
textblob==0.17.1
selenium==4.15.2
webdriver-manager==4.0.0
fake-useragent==1.4.0
